Cast the return value of getTime() to uint64_t to not loose precision.

+++ src/stream/packet.c
@@ -6,7 +6,7 @@ static uint64_t getTime()
 {
 	struct timeval tv;
 	gettimeofday(&tv, 0);
-	return tv.tv_sec * 1000000 + tv.tv_usec;
+	return (uint64_t) tv.tv_sec * 1000000 + tv.tv_usec;
 }
 
 struct yukonPacket *yukonPacketCreate(unsigned char type, unsigned long size)
